requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
azure-functions==1.14.0
    # via -r requirements.in
bdew-datetimes==0.3.10
    # via fristenkalender-generator
fristenkalender-generator==0.0.2
    # via -r requirements.in
holidays==0.26
    # via bdew-datetimes
icalendar==5.0.7
    # via fristenkalender-generator
python-dateutil==2.8.2
    # via
    #   bdew-datetimes
    #   holidays
    #   icalendar
pytz==2023.3
    # via
    #   bdew-datetimes
    #   icalendar
six==1.16.0
    # via python-dateutil
